Completed automating Upload example
